# feat: Optimize package initialization