… data/countrycode_to_region.xml.gz
… data/metadata.xml.gz
… data/prefix_to_carriers/ar.txt.gz
… data/prefix_to_carriers/be.txt.gz
… data/prefix_to_carriers/en.txt.gz
… data/prefix_to_carriers/fa.txt.gz
… data/prefix_to_carriers/ko.txt.gz
… data/prefix_to_carriers/ru.txt.gz
… data/prefix_to_carriers/uk.txt.gz
… data/prefix_to_carriers/zh.txt.gz
… data/prefix_to_carriers/zh_Hant.txt.gz
… data/prefix_to_geocodings/ar.txt.gz
… data/prefix_to_geocodings/be.txt.gz
… data/prefix_to_geocodings/bg.txt.gz
… data/prefix_to_geocodings/bs.txt.gz
… data/prefix_to_geocodings/de.txt.gz
… data/prefix_to_geocodings/el.txt.gz
… data/prefix_to_geocodings/en.txt.gz
… data/prefix_to_geocodings/es.txt.gz
… data/prefix_to_geocodings/fa.txt.gz
… data/prefix_to_geocodings/fi.txt.gz
… data/prefix_to_geocodings/fr.txt.gz
… data/prefix_to_geocodings/hr.txt.gz
… data/prefix_to_geocodings/hu.txt.gz
… data/prefix_to_geocodings/hy.txt.gz
… data/prefix_to_geocodings/id.txt.gz
… data/prefix_to_geocodings/it.txt.gz
… data/prefix_to_geocodings/iw.txt.gz
… data/prefix_to_geocodings/ja.txt.gz
… data/prefix_to_geocodings/ko.txt.gz