This commit improves package startup performance by:

* **Embedding static data:** Replaced go file data loading with `go:embed`, reducing I/O overhead.
* **Optimized metadata build:** Changed `cmd/buildmetadata` to download `metadata` files with the new format.
* **Logarithmic digit counting:** Replaced string conversion and length calculation with a faster logarithmic approach for digit counting.
* **Improve initialization:** Utilized `sync.Map` instead of `map[string]sync.Once` for prevent gunzip the same metadata more than once.
* **Clarify data directory:** Removed the `gen` directory, moving data into the `data` directory because the data is not longer loaded as generated generated code.

These changes result in a substantial reduction in package startup time.

old:
init github.com/nyaruka/phonenumbers @4.2 ms, 7.2 ms clock, 2994832 bytes, 34852 allocs

new:
init github.com/nyaruka/phonenumbers @2.1 ms, 3.7 ms clock, 2917304 bytes, 34444 allocs

the binary size also went down 1M:

4.0K	./main.go
7.8M	./phoneparser-new
8.9M	./phoneparser-main
 17M	.