@rowanseymour 2025-03-19
Have left some comments and generally looks really good... do wish tho the changes were done separately so it'd be possible to see what effect each change is having on package load time.